gate

+++ app/Http/Controllers/UserController.php
@@ -15,12 +15,15 @@ class UserController extends Controller
  
     public function index()
     {
-        return UserResource::collection(User::paginate());
+        $this->authorize('view','users');
+
+        return UserResource::collection(User::with('role')->paginate());
     }
 
   
     public function store(Request $request)
     {
+        $this->authorize('edit','users');
         $user = User::create(
             $request->only('first_name','last_name','email','role_id')
             + ['password' => Hash::make(1234)]
@@ -31,12 +34,14 @@ class UserController extends Controller
     
     public function show($id)
     {
+        $this->authorize('view','users');
         return new UserResource(User::find($id));   
     }
 
    
     public function update(UserUpdateRequest $request, $id)
     {
+        $this->authorize('edit','users');
         $user = User::find($id);
         $user->update($request->only('first_name','last_name','email'));
         return response(new UserResource($user),Response::HTTP_ACCEPTED);
@@ -46,6 +51,7 @@ class UserController extends Controller
   
     public function destroy($id)
     {
+        $this->authorize('edit','users');
         User::destroy($id);
         return response(null,Response::HTTP_NO_CONTENT);
     }
